feat: or-1303 remove brolfeederhoofdactiviteit, move hoofdactiviteit to common project
