+++ src/AssociationRegistry.Admin.Api/Verenigingen/Vereniging.cs
@@ -3,7 +3,8 @@ namespace AssociationRegistry.Admin.Api.Verenigingen;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using AssociationRegistry.Events;
+using AssociationRegistry.Vereniging;
+using Framework;
 using KboNummers;
 using Startdatums;
 using VCodes;
