+++ test/AssociationRegistry.Test/Admin.Api.IntegrationTests/When_posting_a_new_vereniging/Given_A_Valid_Request_With_All_Fields.cs
@@ -3,11 +3,11 @@
 using System.Net;
 using AssociationRegistry.Admin.Api.Constants;
 using AutoFixture;
-using Events;
 using Fixtures;
 using FluentAssertions;
 using Framework.Helpers;
 using Marten;
+using Vereniging;
 using Xunit;
 
 [Collection(VerenigingAdminApiCollection.Name)]
