+++ test/AssociationRegistry.Test/Admin.Api.IntegrationTests/When_posting_a_new_vereniging/Given_A_Valid_Request_With_Minimal_Fields.cs
@@ -2,10 +2,10 @@
 
 using System.Net;
 using AutoFixture;
-using Events;
 using Fixtures;
 using FluentAssertions;
 using Framework.Helpers;
+using Vereniging;
 using Xunit;
 
 [Collection(VerenigingAdminApiCollection.Name)]
