+++ test/AssociationRegistry.Test/Admin.Api.UnitTests/CommandHandlerTests/When_a_RegistreerVerenigingCommand_is_received/Given_A_RegistreerVerenigingCommandHandler.cs
@@ -2,8 +2,8 @@
 
 using AssociationRegistry.Admin.Api.Verenigingen;
 using AutoFixture;
-using Events;
 using FluentAssertions;
+using Vereniging;
 using Xunit;
 
 public class VerenigingRepositoryMock : IVerenigingsRepository
