+++ test/AssociationRegistry.Test/Admin.Api.UnitTests/VerenigingsRepositoryTests/When_saving_a_vereniging/Given_A_New_Vereniging.cs
@@ -4,8 +4,9 @@ using AssociationRegistry.Admin.Api.Events;
 using AssociationRegistry.Admin.Api.Verenigingen;
 using AssociationRegistry.Admin.Api.Verenigingen.VCodes;
 using AssociationRegistry.Admin.Api.Verenigingen.VerenigingsNamen;
-using Events;
+using AssociationRegistry.Framework;
 using FluentAssertions;
+using Vereniging;
 using Xunit;
 
 public class EventStoreMock : IEventStore
