+++ test/AssociationRegistry.Test/Public.Api.IntegrationTests/When_retrieving_a_detail_of_a_vereniging/Given_A_Vereniging_With_Actual_All_Fields.cs
@@ -5,7 +5,7 @@ using Fixtures;
 using FluentAssertions;
 using Xunit;
 using AssociationRegistry.Public.Api.Constants;
-using Events;
+using Vereniging;
 
 public class Given_A_Vereniging_With_Actual_All_Fields_Fixture : PublicApiFixture
 {
